Phenotype predicates and phenotypes are part of `MultiPhenotypeAnalysisResult`.
